Improve user experience for DeleteLanguageButton
If the user tries to delete the currently selected language, give a
warning and let him confirm if he really wants to do that. Also the icon
is greyed out for that language.

+++ lib/widgets/delete_language_button.dart
@@ -5,38 +5,76 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../data/languages.dart';
 
+const ColorFilter greyOut = ColorFilter.matrix(<double>[
+  0.9126,
+  0.7152,
+  0.0722,
+  0,
+  0,
+  0.9126,
+  0.7152,
+  0.0722,
+  0,
+  0,
+  0.9126,
+  0.7152,
+  0.0722,
+  0,
+  0,
+  0,
+  0,
+  0,
+  1,
+  0,
+]);
+
 /// Button to delete one language
 class DeleteLanguageButton extends ConsumerWidget {
   final String languageCode;
-  final bool isDisabled;
-  const DeleteLanguageButton(this.languageCode,
-      {super.key, this.isDisabled = false});
+  const DeleteLanguageButton(this.languageCode, {super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    // discourage deletion the currently selected app language:
+    // - grey the icon out
+    // - if a user still clicks it, let him confirm the deletion first
+    final appLang = ref.watch(appLanguageProvider).languageCode;
+    bool isDiscouraged = (languageCode == appLang);
+
     LanguageController lang = ref.read(languageProvider(languageCode).notifier);
 
-    return IconButton(
-        onPressed: isDisabled
-            ? null
-            : () async {
-//                showDialog(
-//                    context: context, builder: buildPopupDialogCantDelete);
-                // snackbar to be shown after the resources are deleted
-                final snackBar = SnackBar(
-                    content: Text(context.l10n.deletedLanguage(
-                        context.l10n.getLanguageName(languageCode))));
-                await lang.deleteResources();
-                ref.watch(scaffoldMessengerProvider).showSnackBar(snackBar);
-              },
+    final button = IconButton(
+        onPressed: () async {
+          // snackbar to be shown after the resources are deleted
+          final snackBar = SnackBar(
+            content: Text(context.l10n
+                .deletedLanguage(context.l10n.getLanguageName(languageCode))),
+            duration: const Duration(seconds: 1),
+          );
+          if (isDiscouraged) {
+            bool result = await showDialog(
+                context: context,
+                builder: (context) {
+                  return const ConfirmDeletionDialog();
+                });
+            if (!result) return;
+          }
+          await lang.deleteResources();
+          ref.watch(scaffoldMessengerProvider).showSnackBar(snackBar);
+        },
         icon: const Icon(Icons.delete),
         color: Theme.of(context).colorScheme.primary,
         padding: EdgeInsets.zero);
+    if (isDiscouraged) {
+      return ColorFiltered(colorFilter: greyOut, child: button);
+    } else {
+      return button;
+    }
   }
 }
 
 /// Button to delete all languages
-/// (except English and the currently selected app language)
+/// (except the currently selected app language)
 class DeleteAllLanguagesButton extends ConsumerWidget {
   const DeleteAllLanguagesButton({super.key});
 
@@ -50,9 +88,8 @@ class DeleteAllLanguagesButton extends ConsumerWidget {
           int countDeleted = 0;
           String lastLanguage = '';
           for (var languageCode in ref.read(availableLanguagesProvider)) {
-            // Delete all languages except English and the current app language
-            if ((languageCode == 'en') ||
-                (languageCode == ref.read(appLanguageProvider).languageCode)) {
+            // Delete all languages except the current app language
+            if ((languageCode == ref.read(appLanguageProvider).languageCode)) {
               continue;
             }
             if (ref.watch(languageProvider(languageCode)).downloaded) {
@@ -76,5 +113,28 @@ class DeleteAllLanguagesButton extends ConsumerWidget {
         icon: const Icon(Icons.delete),
         color: Theme.of(context).colorScheme.primary,
         padding: EdgeInsets.zero);
+  }
+}
+
+class ConfirmDeletionDialog extends StatelessWidget {
+  const ConfirmDeletionDialog({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+        title: Text(context.l10n.warning),
+        content: Text(context.l10n.warnBeforeDelete),
+        actions: <Widget>[
+          TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(false);
+              },
+              child: Text(context.l10n.cancel)),
+          TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(true);
+              },
+              child: Text(context.l10n.delete))
+        ]);
   }
 }
